# backend/src/startup/swaggerDefs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library Management System API",
    "version": "1.0.0",
    "description": "API documentation for the Library Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/borrow/borrow/{id}": {
      "post": {
        "tags": [
          "Borrow"
        ],
        "summary": "Borrow a book",
        "description": "Borrows a book for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the book to borrow"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dueDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book borrowed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Borrow"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book or User not found"
          },
          "400": {
            "description": "Book not available"
          }
        }
      }
    },
    "/borrow/return/{id}": {
      "post": {
        "tags": [
          "Borrow"
        ],
        "summary": "Return a book",
        "description": "Returns a borrowed book",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the book to return"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Borrow ID"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book, Borrow, or User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Borrow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "book": {
            "type": "string"
          },
          "borrowDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}